Reorganised element types to folders
