+++ assets/src/edit-story/elements/index.js
@@ -6,41 +6,29 @@ import uuid from 'uuid/v4';
 /**
  * Internal dependencies
  */
-import Text from './text';
-import Image from './image';
-import Square from './square';
+import * as textElement from './text';
+import * as imageElement from './image';
+import * as squareElement from './square';
 
-export const createNewElement = ( type, props = {} ) => {
+export const createNewElement = ( type, attributes = {} ) => {
 	const element = elementTypes.find( ( el ) => el.type === type );
-	const defaultProps = element ? element.defaultProps : {};
+	const defaultAttributes = element ? element.defaultAttributes : {};
 	return {
 		type,
 		id: uuid(),
-		...defaultProps,
-		...props,
+		...defaultAttributes,
+		...attributes,
 	};
 };
 
-export const createPage = ( props ) => createNewElement(
-	'page',
-	{
-		elements: [],
-		...props,
-	},
-);
+export const createPage = ( attributes ) => createNewElement( 'page', attributes );
 
 export const elementTypes = [
-	{ type: 'page', defaultProps: { elements: [] }, name: 'Page' },
-	{ type: 'text', defaultProps: Text.defaultProps, component: Text, name: 'Text', panels: Text.panels },
-	{ type: 'image', defaultProps: Image.defaultProps, component: Image, name: 'Image', panels: Image.panels },
-	{ type: 'square', defaultProps: Square.defaultProps, component: Square, name: 'Square', panels: Square.panels },
+	{ type: 'page', defaultAttributes: { elements: [] }, name: 'Page' },
+	{ type: 'text', name: 'Text', ...textElement },
+	{ type: 'image', name: 'Image', ...imageElement },
+	{ type: 'square', name: 'Square', ...squareElement },
 ];
 
 export const getComponentForType =
-	( type ) => elementTypes.find( ( el ) => el.type === type ).component;
-
-export {
-	Text,
-	Image,
-	Square,
-};
+	( type ) => elementTypes.find( ( el ) => el.type === type ).Display;
